proteus/docker/volumes/agent/skills/knowledge-base-management/requirements.txt
# Knowledge Base Management Dependencies

# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0

# Document processing (optional, for full functionality)
# PyPDF2>=3.0.0  # For PDF extraction
# python-docx>=1.1.0  # For DOCX extraction
# beautifulsoup4>=4.12.0  # For HTML parsing

# Database
# (SQLite is included in Python standard library)

# Search (optional, for advanced search)
# faiss-cpu>=1.7.0  # For vector search
# sentence-transformers>=2.2.0  # For embeddings

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Install with: pip install -r requirements.txt
# For minimal install: pip install python-dotenv requests
